+++ R/produce_reference_from_files.R
@@ -8,7 +8,10 @@
 #' @param pi_file Protein interactions from STRING (fully detailed)
 #' @param tf_file Transcription factors
 #' @param tf_column Column containing gene name of TFs
-#' @param bm_file Protein-gene matching from Biomart (N:1)
+#' @param pg_file Protein-gene matching (from Biomart or STRING, N:1)
+#' @param pg_format If STRING it is assumed that the protein column
+#' may have the form *<org_id>.<protein>*. In this case, it will be stripped
+#' before passed to [produce_reference()]
 #' @param out_file Results will be saved to this tab-separated file
 #' @param headers header parameter for [utils::read.table()] for each of the 3 input files
 #' @param seps sep parameter for [utils::read.table()] for each of the 3 input files
@@ -18,8 +21,9 @@
 #' @export
 #'
 #' @examples
-produce_reference_from_files<-function(pi_file,tf_file,tf_column,bm_file,out_file,
-                                       headers=rep(T,3), seps=rep("",3),
+produce_reference_from_files<-function(pi_file,tf_file,tf_column,
+                                       pg_file,pg_format=c("Normal","STRING"),
+                                       out_file,headers=rep(T,3), seps=rep("",3),
                                        ...)
 {
   headers=as.logical(headers)
@@ -31,16 +35,21 @@ produce_reference_from_files<-function(pi_file,tf_file,tf_column,bm_file,out_fil
   else if ((ls=length(seps))<3) seps=c(seps,rep("",3-ls))
 
   # Φόρτωση πρωτεϊνικών αλληλεπιδράσεων
-  t=read.table(pi_file,header = headers[1],sep=seps[1])
+  pi=read.table(pi_file,header = headers[1],sep=seps[1])
 
   # Φόρτωση μεταγραφικών παραγόντων
   tf=read.table(tf_file,header=headers[2],sep=seps[2])
 
-  # Φόρτωση συμπιεσμένου CSV αρχείου biomart
+  # Φόρτωση συμπιεσμένου CSV αρχείου αντιστοίχισης πρωτεϊνών-γονιδίων
   # (περιέχει εγγραφές "Protein stable ID,Gene name")
-  biomart=read.table(bm_file, header=headers[3], sep=seps[3])
+  pg_format=match.arg(pg_format)
+  pg=read.table(pg_file, header=headers[3], sep=seps[3])
+  if (pg_format=="STRING") {
+    pg=pg[,1:2]
+    pg[,1]=gsub("^.*?\\.","",pg[,1])
+  }
 
-  ref=produce_reference(t,tf[,tf_column],biomart,...)
+  ref=produce_reference(pi,tf[,tf_column],pg,...)
 
   # Σώσιμο αποτελεσμάτων
   if (!is.null(ref)) {